reorganised some classes and renamed some names
